I simplified how the calendar works. All events should show up now. The problem seemed to be some bad encoding or something in the json. This should be easier to work on now.
resolves #60
