+++ config/app.php
@@ -172,8 +172,6 @@ return [
         // App\Providers\BroadcastServiceProvider::class,
         App\Providers\EventServiceProvider::class,
         App\Providers\RouteServiceProvider::class,
-
-        CalendarProvider::class,
     ],
 
     /*
