--- resources/views/calendar/html.blade.php
@@ -1,1 +1,0 @@
-<div id='calendar'></div>
