+++ routes/api.php
@@ -13,6 +13,7 @@
 
 Route::middleware('auth:api')->get('/user', 'UserController@me')->name('me');
 
+Route::get('/calendar', 'CalendarController@index');
 Route::get('/orgs', 'apiController@showOrgs')->name('orgs');
 Route::get('/events', 'apiController@showEvents')->name('events');
 Route::get('/homepage/events', 'HomepageController@events')->name('homepage.timeline');
